% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNMFfunction.R
\name{dist_fun}
\alias{dist_fun}
\title{The distance between two matrices of location}
\usage{
dist_fun(X, Y = NULL)
}
\arguments{
\item{X}{matrix of location}

\item{Y}{optional other matrix of location to take distance to}
}
\value{
symmetric matrix of distances.
}
\description{
The distance between two matrices of location
}
