% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNMFfunction.R
\name{dist_index}
\alias{dist_index}
\title{Compute Distance Between Locations}
\usage{
dist_index(X, index)
}
\arguments{
\item{X}{matrix of location}

\item{index}{a vector of the indexes from which you want the distance to all other observations in the location matrix}
}
\value{
matrix of distances between the specified indexes and all other observations in the matrix of locations
}
\description{
The distance between indexes of locations and all other locations in location matrix
}
