% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_lengthscale}
\alias{estimate_lengthscale}
\title{Determine the lengthscale of your data - testing 10 values from 0 until a determined maximum value}
\usage{
estimate_lengthscale(data, location, max_avg_nn = NULL, max_pct = NULL)
}
\arguments{
\item{data}{A matrix including your data}

\item{location}{A matrix with the same number of rows as the data}

\item{max_avg_nn}{An integer determining the the average neighbor for the maximum lengthscale tested}

\item{max_pct}{A number between 0 and 1 determining the percentage of datapoints to the maximum lengthscale}
}
\value{
A data.frame of the length scales and the corresponding test error.
}
\description{
Determine the lengthscale of your data - testing 10 values from 0 until a determined maximum value
}
