% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFbatch.R
\name{groupondist}
\alias{groupondist}
\title{Creates batches of observations that are close in distance}
\usage{
groupondist(location, size = NULL, no_groups = NULL)
}
\arguments{
\item{location}{a matrix of locations (observations x 2)}

\item{size}{an integer determining the size of the batches (overwrites no_group parameter)}

\item{no_groups}{an integer determining the number of batches}
}
\value{
A vector of the same length as the number of locations
}
\description{
Creates batches of observations that are close in distance
}
