\name{init}
\alias{init}
\title{
Initialization strategies for the NMF  based on the k-means
}

\description{
Initialization strategies for the NMF based on the k-means algorithm.
}

\usage{
init(x, k, bs = 1)
}

\arguments{
\item{x}{
An \eqn{n \times D} numerical matrix with data.
}
\item{k}{
The number of lower dimensions. It must be less than the dimensionality of the data, at most \eqn{D-1}.
}
\item{bs}{
The batch size in case the user wants to use the mini-batch k-means algorithm. If bs=1, the classical k-means is used.
}
}

\details{
The function initializes the H matrix for the NMF using the k-means algorithm.
}

\value{
The \eqn{H} matrix, an \eqn{k \times D} matrix.
}

\references{
Sevinc V., Kontemeniotis N., Perdikis T. and Tsagris M. (2026).
Non--negative matrix factorization using the R package nnmf.
https://arxiv.org/pdf/2607.20084

Wang Y. X. and Zhang Y. J. (2012). Nonnegative matrix factorization: A comprehensive review. IEEE Transactions on Knowledge and Data Engineering, 25(6): 1336-1353.

Kim H. and Park H. (2008). Nonnegative matrix factorization based on alternating nonnegativity constrained least squares and active set method. SIAM Journal on Matrix Analysis and Applications, 30(2): 713-730.
}

\author{
Michail Tsagris.

R implementation and documentation: Michail Tsagris \email{mtsagris@uoc.gr}.
}

%\note{
%%  ~~further notes~~
%}

\seealso{
\code{ \link{nmf.qp}
}
}

\examples{
x <- as.matrix(iris[, 1:4])
mod <- nmf.qp(x, 2)
plot(mod$W, colour = iris[, 5])
}

