\name{nmf.qp}
\alias{nmf.qp}
\title{
NMF minimizing the Frobenius norm
}

\description{
NMF minimizing the Frobenius norm using quadratic programming.
}

\usage{
nmf.qp(X, k, H_init = NULL, k_means = TRUE, bs = 1, lr_h = 0.1, tol = 1e-6,
maxiter = 1000, ridge = 1e-8, ncores = 1)
}

\arguments{
\item{X}{
An \eqn{n \times D} numerical matrix with data.
}
\item{k}{
The number of lower dimensions. It must be less than the dimensionality of the data, at most \eqn{D-1}.
}
\item{H_init}{
If you have an initial estimate for H supply it here, otherwise leave it NULL.
}
\item{k_means}{
If this is TRUE, then the K-means algorithm is used to initiate the W and H matrices.
}
\item{bs}{
If you use the K-means algorithm for initialization, you may want to use the mini batch K-means if you have millions of observations. In this case, you need to define the number of batches.
}
\item{lr_h}{
If veo is TRUE, then the exponentiated gradient descent method is used to update the H matrix. In this case you need to supply the value of the learning rate, which is 0.1 by default.
}
\item{tol}{
The tolerance value to terminate the quadratic programming algorithm.
}
\item{maxiter}{
The maximum number of iterations allowed.
}
\item{ridge}{
A small quantity added in the diagonal of the \eqn{D} matrix.
}
\item{ncores}{
Do you want the update of W to be performed in parallel? If yes, specify the number of cores to use.
}
}

\details{
Nonnegative matrix factorization using quadratic programming is performed.
The objective function to be minimized is the square of the Frobenius norm.
If the variables are more than the sample size a hybrid algorithm based on quadratic
programming (for W) and exponentiated gradient descent (for H) is applied.
}

\value{
\item{W}{
The \eqn{W} matrix, an \eqn{n \times k} matrix with the mapped data.
}
\item{H}{
The \eqn{H} matrix, an \eqn{k \times D} matrix.
}
\item{Z}{
The reconstructed data, \eqn{Z = WH}.
}
\item{obj}{
The reconstruction error, \eqn{ ||x - Z||_F^2}.
}
\item{iters}{
The number of iterations performed.
}
\item{runtime}{
The runtime required by the algorithm.
}
}

\references{
Sevinc V., Kontemeniotis N., Perdikis T. and Tsagris M. (2026).
Non--negative matrix factorization using the R package nnmf.
https://arxiv.org/pdf/2607.20084

Wang Y. X. and Zhang Y. J. (2012). Nonnegative matrix factorization: A comprehensive review. IEEE Transactions on Knowledge and Data Engineering, 25(6): 1336-1353.

Kim H. and Park H. (2008). Nonnegative matrix factorization based on alternating nonnegativity constrained least squares and active set method. SIAM Journal on Matrix Analysis and Applications, 30(2): 713-730.
}

\author{
Michail Tsagris, Nikolaos Kontemeniotis and Christos Adam.

R implementation and documentation: Michail Tsagris \email{mtsagris@uoc.gr}, Nikolaos Kontemeniotis \email{kontemeniotisn@gmail.com}
and Christos Adam \email{pada4m4@gmail.com}.
}

%\note{
%%  ~~further notes~~
%}

\seealso{
\code{ \link{nmf.hals}, \link{nmf.sqp}
}
}

\examples{
X <- as.matrix(iris[, 1:4])
mod <- nmf.qp(X, 2)
group <- as.numeric(iris[, 5])
plot(mod$W, col = group)
}

