% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nn_adj}
\alias{nn_adj}
\title{Determine the average nearest neighbors of certain celltypes}
\usage{
nn_adj(location, celltype, nn = 5, sampleid = NULL)
}
\arguments{
\item{location}{A matrix of locations (observation x 2)}

\item{celltype}{A vector of the same length as the number of rows in the location matrix}

\item{nn}{An integer determining the number of nearest neighbors}

\item{sampleid}{A vector to split the location into different samples. Needs to be the same length as the number of rows in the location matrix or NULL (default) if all the observations come from the same sample.}
}
\description{
Determine the average nearest neighbors of certain celltypes
}
