\name{nnmf-package}
\alias{nnmf-package}
\docType{package}
\title{
Nonnegative Matrix Factorization
}

\description{
Nonnegative matrix factorization (NMF) is implemented.
}

\details{
\tabular{ll}{
Package: \tab nnmf\cr
Type: \tab Package\cr
Version: \tab 1.5 \cr
Date: \tab 2026-07-25\cr
License: \tab GPL-2\cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}, Nikolaos Kontemeniotis \email{kontemeniotisn@gmail.com}
and Christos Adam \email{pada4m4@gmail.com}.
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>.
}

\references{
Sevinc V., Kontemeniotis N., Perdikis T. and Tsagris M. (2026).
Non--negative matrix factorization using the R package nnmf.
https://arxiv.org/pdf/2607.20084

Erichson N. B., Mendible A., Wihlborn S. and Kutz J. N. (2018). Randomized nonnegative matrix factorization.
Pattern Recognition Letters, 104: 1-7.

Wang Y. X. and Zhang Y. J. (2012). Nonnegative matrix factorization: A comprehensive review. IEEE Transactions on Knowledge and Data Engineering, 25(6): 1336-1353.

Kim H. and Park H. (2008). Nonnegative matrix factorization based on alternating nonnegativity constrained least squares and active set method. SIAM Journal on Matrix Analysis and Applications, 30(2): 713-730.

Cutler A. and Breiman L. (1994). Archetypal analysis. Technometrics, 36(4): 338--347.
}

%\note{
%}


