% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFbatch.R
\name{nnmf}
\alias{nnmf}
\title{Neighborhood Non-Negative Matrix Factorization}
\usage{
nnmf(
  data,
  noSignatures,
  location,
  lengthscale,
  batch = 1,
  maxiter = 10000,
  tolerance = 1e-08,
  initial = 5,
  smallIter = 100,
  error_freq = 10,
  kernel_cutoff = 0.5,
  normalize = TRUE
)
}
\arguments{
\item{data}{a matrix of non-negative entries with the observations as rows and features as columns}

\item{noSignatures}{integer determining the rank of the factorization, which determines the number of signatures}

\item{location}{a matrix of the spatial locations of the observations}

\item{lengthscale}{the lengthscale of the neighborhood.}

\item{batch}{a vector of the same length as the observations to devide the observations into batches. Default is to have all observations in one batch}

\item{maxiter}{integer determining the maximum number of iterations}

\item{tolerance}{a small value to detmine the stopping criteria. default is 1e-8.}

\item{initial}{}

\item{smallIter}{}

\item{error_freq}{}

\item{kernel_cutoff}{}

\item{normalize}{}
}
\value{
A list of the matrices derived by the factorization and the corresponding generalized Kullback-Leibler
\itemize{
 \item exposures - Non-negative matrix of dimension observations x noSignatures
 \item signatures - Non-negative matrix of dimension noSignatures x features, where rows sum to one.
 \item gkl - Smallest Value of the Generalized Kullback-Leibler
 \item gklvalues - The GKL values at each iteration from error_freq
 }
}
\description{
Finding factors that includes the information of the neighborhood through the spatial information.
}
