% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spatial_data}
\alias{spatial_data}
\title{Construct new dataset of spatial genes}
\usage{
spatial_data(
  data,
  lengthscale,
  location = NULL,
  dist = NULL,
  nfeatures = floor(ncol(data)/2)
)
}
\arguments{
\item{data}{A matrix including your data}

\item{lengthscale}{A double determining the lengthscale. Optimal is to choose the one found from estimate_lengthscale.}

\item{location}{A matrix with the same number of rows as the data}

\item{dist}{A symmetric distance matrix of the data points. Must have the same number of rows and columns as the number of rows in data. If this is specified the location is ignored.}

\item{nfeatures}{An integer determining the number of retrieved spatially variable genes. Default is half of the number of features in the original data}
}
\value{
A list of a new data set containing the most spatially variable genes and the indexes from the original data that is included.
}
\description{
Construct new dataset of spatial genes
}
