% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFbatch.R
\name{topfeatures}
\alias{topfeatures}
\title{Find the top features for each signature}
\usage{
topfeatures(signatures, feature_names, ntop = 10)
}
\arguments{
\item{signatures}{a matrix of the signatures (number of signatures x features)}

\item{feature_names}{a vector of names for the features.}

\item{ntop}{an integer of how many of the top features to output}
}
\value{
a matrix of top features for each signature. (number of signatures x ntop)
}
\description{
Find the top features for each signature
}
