\name{NNMF-package}
\alias{NNMF-package}
\alias{NNMF}
\docType{package}
\title{
\packageTitle{NNMF}
}
\description{
\packageDescription{NNMF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NNMF}
\packageIndices{NNMF}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{NNMF}

Maintainer: \packageMaintainer{NNMF}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
