\name{NNMF-package}
\alias{NNMF-package}
\alias{NNMF}
\docType{package}
\title{
\packageTitle{NNMF}
}
\description{
\packageDescription{NNMF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NNMF}
\packageIndices{NNMF}
The NNMF package provides a fast function 'nnmf' to calculate the factors of non-negative matrix factorization for spatial data. It makes neighboring observations share similar weights for the different factors using gaussian smoothing in the multiplicative updates of NMF. 
}
\author{
\packageAuthor{NNMF}

Maintainer: \packageMaintainer{NNMF}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
