% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNMFfunction.R
\name{nnmf}
\alias{nnmf}
\title{Neighborhood Non-Negative Matrix Factorization}
\usage{
nnmf(
  data,
  noSignatures,
  location = NULL,
  lengthscale = NULL,
  batch = 1,
  maxiter = 1000,
  tolerance = 1e-10,
  initial = 3,
  smallIter = 50,
  error_freq = 10,
  kernel_cutoff = 0.1,
  normalize = TRUE,
  not_sc = FALSE
)
}
\arguments{
\item{data}{a matrix of non-negative entries with the observations as rows and features as columns}

\item{noSignatures}{integer determining the rank of the factorization, which determines the number of signatures}

\item{location}{a matrix of the spatial locations of the observations (observations x 2). If no location is specified regular NMF is used.}

\item{lengthscale}{the lengthscale of the neighborhood. If no lengthscale are specified then the lengthscale is set to 1/10 of the density of your points.}

\item{batch}{a vector of the same length as the observations to divide the observations into batches. Default is to have all observations in one batch.}

\item{maxiter}{integer determining the maximum number of iterations. Default is 10000}

\item{tolerance}{a small value to determine the stopping criteria. Default is 1e-8.}

\item{initial}{the number of intializations. Default is 1.}

\item{smallIter}{the number of iterations to run each initialization. Afterwards the initialization with the smallest error is run till convergence. Default is 100.}

\item{error_freq}{an interger determining how often to calculate the error. Default is 10.}

\item{kernel_cutoff}{a value between 0 and 1, where everything below this value is set to zero in the kernel. Default is 0.5.}

\item{normalize}{a logical value indicating whether the observations should be normalized. Default is TRUE.}

\item{not_sc}{a logical value that indicates if the data is not single cell data(default = TRUE). It is recommended to set to FALSE if the data is not single-cell.}
}
\value{
A list of the matrices derived by the factorization and the corresponding generalized Kullback-Leibler
\itemize{
 \item weights - Non-negative matrix of dimension observations x noSignatures
 \item signatures - Non-negative matrix of dimension noSignatures x features, where rows sum to one.
 \item error - Final error of the Generalized Kullback-Leibler
 \item errorvalues - a vector of the length of maxiter. Includes the GKL values calculated at the frequency of error_freq
 \item avg_nn - The average number of nearest neighbors included in averaging.
 \item lengthscale - the lengthscale used for the analysis
 \item batch - The batches of the analysis, which is 1 if all were in one batch. 
 }
}
\description{
Finding factors that includes the information of the neighborhood through the spatial information.
}
